Split BDEW module
