+++ demandlib/bdew/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+Load profiles
+
+SPDX-FileCopyrightText: Patrik Schönfeldt
+
+SPDX-License-Identifier: MIT
+"""
+
+from .elec_slp import ElecSlp
+from .heat_building import HeatBuilding
