+++ demandlib/bdew/bdew_data/shlp_weekday_factors.csv
